+++ internal/repo/storage.go
@@ -1,6 +1,6 @@
 package repo
 
 type LineStorage interface {
-	Save(coef float32) error
+	Save(coef float64) error
 	GetAll() ([]CoefItem, error)
 }
